(TK-385) Test overflow events handling at the clojurize unit level
Previously we tested the handling of overflow events as a unit test on
`process-events!`.

As part of this PR we consolidate the handling of event representations
to the function `clojurize` and move the call(s) to `clojurize` outside
of the `process-events!` function. It is therefore no longer appropriate
to test our overflow handling with `process-events!` but the handling
can be better tested with `clojurize`.

This refactors the overflow test to test `clojurize` rather than
`process-events!`.